+++ src/main/java/janggi/manager/JanggiGame.java
@@ -1,4 +1,7 @@
-package janggi;
+package janggi.manager;
+
+import janggi.view.InputView;
+import janggi.view.OutputView;
 
 public class JanggiGame {
